Added setting option to specify a providers directory.
This can be used to extend the functionality of the client with new
providers without cluttering the core 'providers' directory, as well as
maintaining custom providers between client installs/upgrades.

Providers in this directory will be loaded after the core providers, and
so may be used to replace the core providers.