Added command line arguments using Yargs.
Any arguments passed to the client will be included in the environment
object as 'args'.

--settingsFile
Used to pass the path to a custom settings file that the client should
use.

--token
Used to pass a token directly on the command line.

--tokenFile
Used to pass a path to a file containing the client token.

The token options can be used to overwrite an existing token by passing
the --force parameter.